# path{,/filepath}: speed up Match